requirements.txt
aiofiles==0.5.0
argon2-cffi==20.1.0
astropy==4.0.2
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
blinker==1.4
cffi==1.14.3
click==7.1.2
decorator==4.4.2
defusedxml==0.6.0
dictdiffer==0.8.1
entrypoints==0.3
ephem==3.7.7.1
flake8==3.8.4
h11==0.11.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.1
hyperframe==6.0.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.7
nbformat==5.0.7
nest-asyncio==1.4.1
notebook==6.1.4
nptyping==1.3.0
numpy==1.19.2
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
priority==1.3.0
prometheus-client==0.8.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.4
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
Quart==0.13.1
rope==0.18.0
Send2Trash==1.5.0
six==1.15.0
terminado==0.9.1
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4
typish==1.9.1
watchgod==0.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wsproto==0.15.0
